Update to use hollowstrawberry/kohya-colab
This is a total rework of the image dependencies to accommodate running
the newer Jupyter Lab instead of legacy Jupyter Notebook, as well as a
better colab adaptation that supports min-snr-gamma for efficient
training, much better results, and much simpler interactive process.

This commit is still missing the modified colab ipynb. For now this is
left as an exercise for the reader, but I plan to add one soon.